Responding to comments on PR from copilot.  Reorganizing where the logic for the system information and channel information is passed
